Rename prober2 to prober
