+++ pkg/prober/prober_manager_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2019-2021 VMware, Inc. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package prober2
+package prober
 
 import (
 	goctx "context"
@@ -18,10 +18,10 @@ import (
 
 	vmopv1 "github.com/vmware-tanzu/vm-operator/api/v1alpha2"
 
-	"github.com/vmware-tanzu/vm-operator/pkg/prober2/context"
-	fakeworker "github.com/vmware-tanzu/vm-operator/pkg/prober2/fake/worker"
-	"github.com/vmware-tanzu/vm-operator/pkg/prober2/probe"
-	"github.com/vmware-tanzu/vm-operator/pkg/prober2/worker"
+	"github.com/vmware-tanzu/vm-operator/pkg/prober/context"
+	fakeworker "github.com/vmware-tanzu/vm-operator/pkg/prober/fake/worker"
+	"github.com/vmware-tanzu/vm-operator/pkg/prober/probe"
+	"github.com/vmware-tanzu/vm-operator/pkg/prober/worker"
 	"github.com/vmware-tanzu/vm-operator/pkg/record"
 	"github.com/vmware-tanzu/vm-operator/pkg/vmprovider/fake"
 	"github.com/vmware-tanzu/vm-operator/test/builder"
