+++ pkg/prober/worker/prober_worker.go
@@ -8,8 +8,8 @@ import (
 
 	vmopv1 "github.com/vmware-tanzu/vm-operator/api/v1alpha2"
 
-	"github.com/vmware-tanzu/vm-operator/pkg/prober2/context"
-	"github.com/vmware-tanzu/vm-operator/pkg/prober2/probe"
+	"github.com/vmware-tanzu/vm-operator/pkg/prober/context"
+	"github.com/vmware-tanzu/vm-operator/pkg/prober/probe"
 )
 
 // Worker represents a prober worker interface.
